App/requirements.txt
streamlit
torch
torchvision
grad-cam
opencv-python-headless
Pillow